vaults move: ../data_sources/[dsId]/tables/[tId]/rows/[rId]
